Fix list-all for Linuxbrew
There is a known issue with Linuxbrew which was breaking the list-all
script when used through asdf (e.g asdf list-all R). There was a fix
already for this issue (see https://github.com/asdf-community/asdf-r/pull/6).

What I did here was to return the versions on the same format they were
returned before so asdf knows what to do with them.